requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
APScheduler==3.10.4